added ui for user fav page
