+++ frontend/src/Components/NavBar/NavBar.css
@@ -1,5 +1,5 @@
 .nav-bar {
-  background-color: #fff;
+  background-color: #ff5f6d;
   padding: 10px 20px;
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
   position: fixed;
@@ -29,6 +29,7 @@
   font-weight: bold;
   padding: 5px 10px;
   transition: color 0.3s;
+  font-size: 1em;
 }
 
 .nav-link::before {
