+++ frontend/src/Components/ProviderHomePage/ProviderHomePage.css
@@ -1,0 +1,163 @@
+body,
+html {
+  margin: 0;
+  padding: 0;
+  font-family: Arial, sans-serif;
+  background-color: #f0f2f5;
+}
+
+.homepage-header {
+  text-align: center;
+  margin: 150px 0;
+  color: #333;
+}
+
+.service-provider-homepage {
+  padding: 20px;
+  width: 1200px;
+  margin: 0 auto;
+  background-color: #ffffff;
+  border-radius: 10px;
+  box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1);
+}
+
+.homepage-header h1 {
+  color: #ff5f6d;
+  text-align: center;
+}
+section {
+  background: white;
+  border: 1px solid #ddd;
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  padding: 20px;
+  margin-bottom: 20px;
+}
+
+section h2 {
+  margin-top: 0;
+  color: #333;
+}
+
+section p {
+  color: #555;
+}
+table {
+  width: 100%;
+  border-collapse: collapse;
+  margin: 20px 0;
+}
+
+table th,
+table td {
+  padding: 10px;
+  text-align: left;
+  border-bottom: 1px solid #ddd;
+}
+
+table th {
+  background-color: #ff5f6d;
+  color: white;
+}
+
+button {
+  display: inline-block;
+  padding: 10px 20px;
+  margin: 10px 0;
+  font-size: 1em;
+  font-weight: bold;
+  color: white;
+  background-color: #ff5f6d;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: background-color 0.3s;
+}
+
+button:hover {
+  background-color: #e64552;
+}
+
+.media-gallery {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+}
+
+.media-title {
+  width: 100%;
+  font-size: 1.2em;
+  margin-top: 20px;
+  color: #555;
+}
+
+.media-section {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+}
+
+.media-item img,
+.media-item video {
+  max-width: 200px;
+  border-radius: 10px;
+  box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1);
+}
+
+.error-message {
+  color: red;
+  text-align: center;
+  margin: 10px 0;
+}
+
+/* Schedule section */
+.schedule-section ul {
+  list-style-type: none;
+  padding: 0;
+}
+
+.schedule-section li {
+  padding: 10px;
+  border-bottom: 1px solid #ddd;
+}
+
+.schedule-section li strong {
+  color: #333;
+}
+
+.schedule-section button {
+  margin-top: 20px;
+}
+
+/* Appointments section */
+.appointments-section table th,
+.appointments-section table td {
+  border: 1px solid #ddd;
+  padding: 10px;
+}
+
+.appointments-section table th {
+  background-color: #ff5f6d;
+  color: white;
+}
+
+.appointments-section table td {
+  background-color: #ffffff;
+}
+
+.logout-button {
+  background-color: white;
+  color: #ff5f6d;
+  border: none;
+  margin: 0, 90px;
+  padding: 10px 30px;
+  border-radius: 5px;
+  font-weight: bold;
+  cursor: pointer;
+  transition: background-color 0.3s;
+}
+
+.logout-button:hover {
+  background-color: #e64552;
+  color: white;
+}
