+++ frontend/src/Components/SelectProfileModal/SelectProfileModal.css
@@ -2,62 +2,63 @@
   position: fixed;
   top: 0;
   left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.5);
+  width: 100%;
+  height: 100%;
+  background-color: lightgray;
   display: flex;
   justify-content: center;
   align-items: center;
+  z-index: 1050;
 }
 
 .modal-content {
-  background-color: rgb(240, 236, 236);
+  background: white;
   padding: 20px;
-  border-radius: 10px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  width: 400px;
-  height: 360px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.modal-header{
+  border-radius: 19px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+  width: 90%;
+  max-width: 400px;
+  height: 400px;
   text-align: center;
-  text-decoration: underline;
-  font-weight: bold;
 }
 
-.user-button,
-.provider-button {
-  margin-bottom: 40px;
-  padding: 20px 20px;
-  cursor: pointer;
+.modal-header {
+  margin-top: 0;
+  color: rgb(8, 8, 8);
+}
+
+button {
+  padding: 10px 20px;
+  margin-top: 20px;
   border: none;
   border-radius: 5px;
-  font-size: 16px;
-  color: #fff;
   font-weight: bold;
-  background-color: #ff8400;
+  cursor: pointer;
+  transition: background-color 0.3s;
 }
 
-.user-button:hover,
-.provider-button:hover {
-  background-color: #ff9900;
+.user-button {
+  background-color: #ff5f6d;
+  color: white;
+  border-radius: 70px;
+  margin-top: 40px;
+}
+
+.provider-button {
+  background-color: #ff5f6d;
+  margin-top: 80px;
+  color: white;
+  border-radius: 70px;
 }
 
 .close-button {
-  margin-top: auto;
-  padding: 10px;
-  width: 20%;
-  cursor: pointer;
-  border: none;
-  border-radius: 5px;
-  font-size: 16px;
-  color: #fff;
-  background-color: #dc3545;
+  margin-top: 80px;
+  background-color: #777;
+  color: white;
+  width: 100px;
+  border-radius: 10px;
 }
 
-.close-button:hover {
-  background-color: #c82333;
+button:hover {
+  opacity: 0.9;
 }
