+++ frontend/src/Components/SelectProfilePage/SelectProfilePage.css
@@ -24,8 +24,7 @@
   border: none;
   border-radius: 5px;
   font-size: 16px;
-  color: #fff;
+  color: white;
   font-weight: bold;
-  background-color: #ff8400;
-
+  background-color: #ff5f6d;
 }
