+++ frontend/src/Components/ServiceProviderModal/ServiceProviderModal.css
@@ -1,5 +1,5 @@
 .modal-dialog {
-    max-width: 800px;
+    max-width: 500px;
     margin: 1.75rem auto;
 }
 
@@ -53,3 +53,4 @@
 .modal-footer .btn {
     margin-left: 0.5rem;
 }
+
