+++ frontend/src/Components/UserHomePage/UserHomePage.jsx
@@ -55,7 +55,9 @@ const UserHomePage = () => {
       </div>
       <div className = "home-container">
       <div className="category-section braid-section" onClick={() => handleCategoryClick('Braids')}>
-        <div className="tooltip">Discover beautiful braids styles</div> BRAIDS</div>
+        <div className="tooltip">
+          Discover beautiful braids styles
+        </div> BRAIDS</div>
       <div className="category-section haircut-section" onClick={() => handleCategoryClick('Haircuts')}>
         <div className="tooltip">Explore stylish haircuts</div>HAIRCUTS</div>
       <div className="category-section weave-section" onClick={() => handleCategoryClick('Weave and Installs')}>
@@ -65,7 +67,7 @@ const UserHomePage = () => {
       </div>
       {errorMessage && <p className="error-message">{errorMessage}</p>}
       <section>
-        <h2>Upcoming Bookings</h2>
+        <h2>Upcoming Appointments</h2>
         {appointments.length > 0 ? (
           appointments.map(appointment => (
             <div key={appointment.id} className="appointment-card">
